final version

+++ calculator.py
@@ -9,7 +9,8 @@ def prompt(message):
 
 def invalid_number(number_str):
     try:
-        int(number_str)
+        float(number_str)
+    
     except ValueError:
         return True
 
@@ -37,16 +38,16 @@ def calculator():
     while operation not in ["1", "2", "3", "4"]:
         prompt(MESSAEGS['invalid_operation'][language])
         operation = input()
-
+    
     match operation:
         case "1":
-            output = int(number1) + int(number2)
+            output = float(number1) + float(number2)
         case "2":
-            output = int(number1) - int(number2)
+            output = float(number1) - float(number2)
         case "3":
-            output = int(number1) * int(number2)
+            output = float(number1) * float(number2)
         case "4":
-            output = int(number1) / int(number2)
+            output = float(number1) / float(number2)
 
     prompt(f"The result is {output}")
